pyproject.toml
[tool.poetry]
name = "restate-prototype"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.13"
restate_sdk = "^0.5.1"
fastapi = "^0.110.0"
uvicorn = "^0.27.0"
pydantic = "^2.6.0"
strawberry-graphql = "^0.219.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
